server/requirements.txt
# Aerospike dependencies
aerospike-vector-search

# Web Server 
fastapi
uvicorn[standard]
python-multipart

# Web Scraper
#scrapy
#scrapy-playwright
#markdownify

# RAG dependencies
openai
transformers
torch
einops
google-cloud-aiplatform
sentence-transformers==2.2.2
